// frontend/shared/src/graphql/generated/FreetextAnswerFragment.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL fragment: FreetextAnswerFragment
// ====================================================

export interface FreetextAnswerFragment {
  __typename: "FreetextAnswer";
  surveyInvitationId: string;
  questionnaireId: string;
  questionId: string;
  text: string;
}